refactor: rename AutoApp to TigrblApp
