+++ pkgs/standards/tigrbl/tests/i9n/test_rest_row_serialization.py
@@ -5,7 +5,7 @@ from httpx import ASGITransport, AsyncClient
 from sqlalchemy import Integer, String, select
 from sqlalchemy.orm import Mapped
 
-from tigrbl.v3.autoapp import AutoApp as Tigrblv3
+from tigrbl.v3 import TigrblApp as Tigrblv3
 from tigrbl.v3.engine.shortcuts import mem
 from tigrbl.v3.specs import F, IO, S, acol
 from tigrbl.v3.orm.tables import Base as Base3
