+++ pkgs/standards/tigrbl/tests/i9n/test_row_result_serialization.py
@@ -5,7 +5,7 @@ from httpx import ASGITransport, AsyncClient
 from sqlalchemy import Integer, String, select
 from sqlalchemy.orm import Mapped
 
-from tigrbl.v3.autoapp import AutoApp as Tigrblv3
+from tigrbl.v3 import TigrblApp as Tigrblv3
 from tigrbl.v3.engine import resolver as _resolver
 from tigrbl.v3.engine.shortcuts import mem
 from tigrbl.v3.specs import S, acol
