+++ pkgs/standards/tigrbl/tests/unit/test_app_reexport.py
@@ -1,6 +1,6 @@
 from tigrbl.v3 import App
-from tigrbl.v3.deps.fastapi import App as FastAPIApp
+from tigrbl.v3.app._app import App as InternalApp
 
 
 def test_app_reexport():
-    assert App is FastAPIApp
+    assert App is InternalApp
