+++ pkgs/standards/tigrbl/tests/unit/test_include_models_base_prefix.py
@@ -1,6 +1,6 @@
 from tigrbl.v3.types import App
 
-from tigrbl.v3.autoapp import AutoApp
+from tigrbl.v3 import TigrblApp
 from tigrbl.v3.orm.tables import Base
 from tigrbl.v3.orm.mixins import GUIDPk
 from tigrbl.v3.types import Column, String
@@ -17,7 +17,7 @@ def test_include_models_base_prefix_avoids_duplicate_segments():
         __tablename__ = "key_versions"
         name = Column(String, nullable=False)
 
-    api = AutoApp()
+    api = TigrblApp()
     api.include_models([Key, KeyVersion], base_prefix="/kms")
     app.include_router(api.router)
