+++ pkgs/standards/tigrbl/tests/unit/test_initialize_cross_ddl.py
@@ -1,5 +1,5 @@
 import pytest
-from tigrbl.v3.autoapp import AutoApp
+from tigrbl.v3 import TigrblApp
 from tigrbl.v3 import Base
 from tigrbl.v3.specs import S, acol
 from tigrbl.v3.engine.shortcuts import mem
@@ -20,7 +20,7 @@ def _model():
 async def test_initialize_with_sync_engine():
     Base.metadata.clear()
     Widget = _model()
-    api = AutoApp(engine=mem(async_=False))
+    api = TigrblApp(engine=mem(async_=False))
     api.include_model(Widget)
     await api.initialize()
     assert getattr(api.tables, "Widget").name == "widgets"
