+++ pkgs/standards/tigrbl/tests/unit/test_rest_no_schema_jsonable.py
@@ -3,7 +3,7 @@ from fastapi.testclient import TestClient
 from sqlalchemy import Integer, String
 from sqlalchemy.orm import Mapped
 
-from tigrbl.v3.autoapp import AutoApp as Tigrblv3
+from tigrbl.v3 import TigrblApp as Tigrblv3
 from tigrbl.v3.engine.shortcuts import mem
 from tigrbl.v3.specs import F, IO, S, acol
 from tigrbl.v3.orm.tables import Base as Base3
