+++ pkgs/standards/tigrbl/tests/unit/test_rest_rpc_symmetry.py
@@ -1,6 +1,6 @@
 import inspect
 
-from tigrbl.v3 import AutoApp
+from tigrbl.v3 import TigrblApp
 from tigrbl.v3.orm.tables import Base
 from tigrbl.v3.orm.mixins import (
     GUIDPk,
@@ -24,7 +24,7 @@ def _rest_param_names(route):
 
 
 def _assert_symmetry(model, verbs, rest_params):
-    api = AutoApp()
+    api = TigrblApp()
     api.include_model(model, mount_router=False)
 
     expected_verbs = set(verbs)
