+++ pkgs/standards/tigrbl/tests/unit/test_sqlite_attachments.py
@@ -1,5 +1,5 @@
 import pytest
-from tigrbl.v3.autoapp import AutoApp
+from tigrbl.v3 import TigrblApp
 
 
 def _db_names(conn):
@@ -9,7 +9,7 @@ def _db_names(conn):
 
 def test_initialize_sync_without_sqlite_attachments(sync_db_session):
     engine, get_db = sync_db_session
-    api = AutoApp(get_db=get_db)
+    api = TigrblApp(get_db=get_db)
     api.initialize()
     with engine.connect() as conn:
         assert _db_names(conn) == {"main"}
@@ -21,7 +21,7 @@ def test_initialize_sync_with_sqlite_attachments(sync_db_session, tmp_path):
     engine, get_db = sync_db_session
     attach_db = tmp_path / "logs.sqlite"
     attach_db.touch()
-    api = AutoApp(get_db=get_db)
+    api = TigrblApp(get_db=get_db)
     api.initialize(sqlite_attachments={"logs": str(attach_db)})
     with engine.connect() as conn:
         assert "logs" in _db_names(conn)
@@ -32,7 +32,7 @@ def test_initialize_sync_with_sqlite_attachments(sync_db_session, tmp_path):
 @pytest.mark.asyncio
 async def test_initialize_async_without_sqlite_attachments(async_db_session):
     engine, get_db = async_db_session
-    api = AutoApp(get_db=get_db)
+    api = TigrblApp(get_db=get_db)
     await api.initialize()
     async with engine.connect() as conn:
         result = await conn.exec_driver_sql("PRAGMA database_list")
@@ -47,7 +47,7 @@ async def test_initialize_async_with_sqlite_attachments(async_db_session, tmp_pa
     engine, get_db = async_db_session
     attach_db = tmp_path / "logs.sqlite"
     attach_db.touch()
-    api = AutoApp(get_db=get_db)
+    api = TigrblApp(get_db=get_db)
     await api.initialize(sqlite_attachments={"logs": str(attach_db)})
     async with engine.connect() as conn:
         result = await conn.exec_driver_sql("PRAGMA database_list")
